Fix args order and tests
